requirements.txt
paramiko
pyyaml
pyinstaller
